refactor: reorganise subgraph_service.rs into a services/subgraph/ module tree
Splits the flat subgraph.rs and subgraph_service.rs files into services/subgraph/{mod,types,service}.rs plus an http/ submodule for HTTP client utilities (TLS config, response parsing), so individual concerns can be extracted into Tower layers going forward. All previously accessible paths under services::subgraph and services::subgraph_service are preserved via re-exports.
